Removed duplicate deploy logic
