+++ packages/contracts/src/deployment/index.ts
@@ -1,1 +1,2 @@
 export * from './contract-deploy'
+export { AddressResolverMapping, RollupDeployConfig } from './types'
