+++ packages/contracts/test/test-helpers/index.ts
@@ -6,5 +6,5 @@ export * from './ethereum-helpers'
 export * from './ovm-helpers'
 export * from './trie-helpers'
 export * from './wallet-helpers'
-export * from './resolution'
 export * from './execution-manager-helpers'
+export * from './deployment-helpers'
