--- packages/contracts/test/test-helpers/resolution/index.ts
@@ -1,2 +1,0 @@
-export * from './resolver'
-export { AddressResolverMapping } from './types'
