+++ packages/contracts/test/test-helpers/types/deployment.ts
@@ -1,0 +1,1 @@
+export { AddressResolverMapping } from '../../../src'
