+++ packages/contracts/test/test-helpers/types/index.ts
@@ -1,2 +1,3 @@
 export * from './batch'
 export * from './convenience'
+export * from './deployment'
